Add virtio-rng device to default configuration
virtio-rng provides a way for the host to provide entropy to the guest.
VM guests generally have very few sources of quality entropy, which is why
tools like haveged and rngd exist to provide a last resort provider of
entropy.

The source of entropy is configurable and defaults to /dev/urandom. Ideally
linux would provide a random device that blocked only at startup and then
never again, but that does not exist.